made the job to ran on Hadoop and to fetch results
