+++ monitor/services/src/test/java/org/datacleaner/monitor/server/job/DefaultJobEngineManagerTest.java
@@ -59,9 +59,6 @@ public class DefaultJobEngineManagerTest extends TestCase {
         
         engine = manager.getJobEngine(CustomJobContext.class);
         assertEquals(CustomJobEngine.class, engine.getClass());
-        
-        engine = manager.getJobEngineOfType(HadoopDataCleanerJobEngine.class); 
-        assertNotNull(engine);
     }
     
     public void testGetJobEngineHierarchy() throws Exception {
@@ -78,10 +75,5 @@ public class DefaultJobEngineManagerTest extends TestCase {
 
         engine = manager.getJobEngine(MetricJobContext.class);
         assertNull(engine);
-        
-        engine = manager.getJobEngineOfType(HadoopDataCleanerJobEngine.class); 
-        assertNotNull(engine);
-        
-                
     }
 }
