+++ monitor/services/src/test/resources/example_hadoop_repo/tenant/conf.xml
@@ -1,0 +1,161 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<configuration xmlns="http://eobjects.org/analyzerbeans/configuration/1.0">
+
+	<datastore-catalog>
+		<jdbc-datastore description="DataCleaner sample database, containing an example data warehouse with orders, customers, products, employees etc." name="orderdb">
+			<url>jdbc:hsqldb:res:orderdb</url>
+			<driver>org.hsqldb.jdbcDriver</driver>
+			<username>SA</username>
+			<password/>
+		</jdbc-datastore>
+		
+		<csv-datastore description="A growing list of known product vendor names, used as a reference dictionary in the example product profiling job to ensure that vendor names are correctly filled and spelled." name="Vendor list">
+			<filename>reference-data/vendor_dictionary.txt</filename>
+		</csv-datastore>
+
+		<pojo-datastore description="A simple example datastore with a (limited) selection of contributors to the DataCleaner open source project." name="Contributors">
+			<schema-name>contributors</schema-name>
+			<table>
+				<name>Companies</name>
+				<columns>
+					<column>
+						<name>ID</name>
+						<type>INTEGER</type>
+					</column>
+					<column>
+						<name>Name</name>
+						<type>VARCHAR</type>
+					</column>
+				</columns>
+				<rows>
+					<row>
+						<v>1</v>
+						<v>Human Inference</v>
+					</row>
+					<row>
+						<v>2</v>
+						<v>Xebia</v>
+					</row>
+				</rows>
+			</table>
+			<table>
+				<name>Persons</name>
+				<columns>
+					<column>
+						<name>ID</name>
+						<type>VARCHAR</type>
+					</column>
+					<column>
+						<name>Name</name>
+						<type>VARCHAR</type>
+					</column>
+					<column>
+						<name>Email</name>
+						<type>LIST</type>
+					</column>
+					<column>
+						<name>Country</name>
+						<type>VARCHAR</type>
+					</column>
+					<column>
+						<name>Twitter account</name>
+						<type>VARCHAR</type>
+					</column>
+					<column>
+						<name>Company ID</name>
+						<type>INTEGER</type>
+					</column>
+				</columns>
+				<rows>
+					<row>
+						<v>kasper</v>
+						<v>Kasper Sørensen</v>
+						<v>[kasper@eobjects.org,kasper.sorensen@humaninference.com,i.am.kasper.sorensen@gmail.com]</v>
+						<v>Denmark</v>
+						<v>kaspersor</v>
+						<v>1</v>
+					</row>
+					<row>
+						<v>stefan</v>
+						<v>Stefan Janssen</v>
+						<v>[stefan.janssen@humaninference.com]</v>
+						<v>The Netherlands</v>
+						<v/>
+						<v>1</v>
+					</row>
+					<row>
+						<v>ankitk</v>
+						<v>Ankit Kumar</v>
+						<v>[ankit.kumar@humaninference.com]</v>
+						<v>The Netherlands</v>
+						<v/>
+						<v>1</v>
+					</row>
+					<row>
+						<v>nancysharma</v>
+						<v>Nancy Sharma</v>
+						<v>[nsharma@xebia.com]</v>
+						<v>India</v>
+						<v/>
+						<v>2</v>
+					</row>
+					<row>
+						<v>paritoshranjan</v>
+						<v>Paritosh Ranjan</v>
+						<v>[pranjan@xebia.com]</v>
+						<v>India</v>
+						<v>paritoshranjan</v>
+						<v>2</v>
+					</row>
+					<row>
+						<v>rajneesh</v>
+						<v>Rajneesh Namta</v>
+						<v>[rnamta@xebia.com]</v>
+						<v>India</v>
+						<v>rnamta</v>
+						<v>2</v>
+					</row>
+					<row>
+						<v>swapnil.dubey</v>
+						<v>Swapnil Dubey</v>
+						<v>[sdubey@xebia.com]</v>
+						<v>India</v>
+						<v/>
+						<v>2</v>
+					</row>
+					<row>
+						<v>nikhil</v>
+						<v>Nikhil Srivastava</v>
+						<v>[nsrivastava@xebia.com]</v>
+						<v>India</v>
+						<v/>
+						<v>2</v>
+					</row>
+				</rows>
+			</table>
+		</pojo-datastore>
+		
+		  <csv-datastore name="customers.csv">
+      <filename>hdfs://{org.datacleaner.hadoop.environment}/datacleaner/customers.csv</filename>
+      <quote-char>"</quote-char>
+      <separator-char>,</separator-char>
+      <escape-char>NOT_A_CHAR</escape-char>
+      <encoding>UTF-8</encoding>
+      <fail-on-inconsistencies>false</fail-on-inconsistencies>
+      <multiline-values>false</multiline-values>
+      <header-line-number>1</header-line-number>
+    </csv-datastore>
+	</datastore-catalog>
+
+	<reference-data-catalog>
+		<dictionaries>
+			<datastore-dictionary name="vendor whitelist">
+				<datastore-name>Vendor list</datastore-name>
+				<column-path>VENDOR</column-path>
+			</datastore-dictionary>
+		</dictionaries>
+		<synonym-catalogs>
+		</synonym-catalogs>
+	</reference-data-catalog>
+
+</configuration>
