+++ monitor/ui/pom.xml
@@ -1,4 +1,5 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.eobjects.datacleaner</groupId>
@@ -15,7 +16,8 @@
 				<artifactId>gwt-maven-plugin</artifactId>
 				<version>${gwt.version}</version>
 				<configuration>
-					<!-- GWT uses ASM v5 and we have v4 on classpath. This flag avoids the conflict during GWT compile -->
+					<!-- GWT uses ASM v5 and we have v4 on classpath. This flag avoids the 
+						conflict during GWT compile -->
 					<gwtSdkFirstInClasspath>true</gwtSdkFirstInClasspath>
 				</configuration>
 				<executions>
@@ -176,6 +178,22 @@
 			<artifactId>DataCleaner-monitor-documentation</artifactId>
 			<version>${project.version}</version>
 		</dependency>
+		<!--Hadoop dependencies -->
+		<dependency>
+			<groupId>org.apache.hadoop</groupId>
+			<artifactId>hadoop-hdfs</artifactId>
+			<exclusions>
+				<exclusion>
+					<groupId>log4j</groupId>
+					<artifactId>log4j</artifactId>
+				</exclusion>
+			</exclusions>
+		</dependency>
+
+		<dependency>
+			<groupId>org.apache.spark</groupId>
+			<artifactId>spark-launcher_2.11</artifactId>
+		</dependency>
 
 		<!-- DataCleaner monitor extensions -->
 		<dependency>
@@ -271,7 +289,7 @@
 			<version>2.2</version>
 			<scope>provided</scope>
 		</dependency>
-		
+
 		<!-- Pretty faces - URL rewriting for JSF 2.x -->
 		<dependency>
 			<groupId>org.ocpsoft.rewrite</groupId>
